[#3250] fix failures on macos -> cast to (uint64_t)

+++ nes-runtime/src/Execution/Aggregation/AvgAggregation.cpp
@@ -15,7 +15,6 @@
 #include <Execution/Aggregation/AvgAggregation.hpp>
 #include <Nautilus/Interface/DataTypes/Integer/Int.hpp>
 #include <Execution/Aggregation/AggregationValue.hpp>
-#include <cstddef>
 
 namespace NES::Runtime::Execution::Aggregation {
 
@@ -26,7 +25,7 @@ void AvgAggregationFunction::lift(Nautilus::Value<Nautilus::MemRef> memref, Naut
     // load memref
     auto oldSum = memref.load<Nautilus::Int64>();// TODO 3280 check the type
     // calc the offset to get Memref of the count value
-    auto countMemref = (memref+offsetof(AvgAggregationValue, count)).as<Nautilus::MemRef>();
+    auto countMemref = (memref+ (uint64_t) offsetof(AvgAggregationValue, count)).as<Nautilus::MemRef>();
     auto oldCount = countMemref.load<Nautilus::Int64>();
 
     // add the values
@@ -41,12 +40,12 @@ void AvgAggregationFunction::combine(Nautilus::Value<Nautilus::MemRef> memref1, 
     // load memref1
     auto sumLeft = memref1.load<Nautilus::Int64>();
     // calc the offset to get Memref of the count value
-    auto countLeftMemref = (memref1+offsetof(AvgAggregationValue, count)).as<Nautilus::MemRef>();
+    auto countLeftMemref = (memref1+ (uint64_t) offsetof(AvgAggregationValue, count)).as<Nautilus::MemRef>();
     auto countLeft = countLeftMemref.load<Nautilus::Int64>();
     // load memref2
     auto sumRight = memref2.load<Nautilus::Int64>();
     // calc the offset to get Memref of the count value
-    auto countRightMemref = (memref2+offsetof(AvgAggregationValue, count)).as<Nautilus::MemRef>();
+    auto countRightMemref = (memref2+ (uint64_t) offsetof(AvgAggregationValue, count)).as<Nautilus::MemRef>();
     auto countRight = countRightMemref.load<Nautilus::Int64>();
 
     // add the values
@@ -60,7 +59,7 @@ void AvgAggregationFunction::combine(Nautilus::Value<Nautilus::MemRef> memref1, 
 Nautilus::Value<> AvgAggregationFunction::lower(Nautilus::Value<Nautilus::MemRef> memref) {
     // load memrefs
     auto sum = memref.load<Nautilus::Int64>();// TODO 3280 check the type
-    auto countMemref = (memref+offsetof(AvgAggregationValue, count)).as<Nautilus::MemRef>();
+    auto countMemref = (memref+ (uint64_t) offsetof(AvgAggregationValue, count)).as<Nautilus::MemRef>();
     auto count = countMemref.load<Nautilus::Int64>();
     // calc the average
     auto finalVal = sum / count;
@@ -70,7 +69,7 @@ Nautilus::Value<> AvgAggregationFunction::lower(Nautilus::Value<Nautilus::MemRef
 
 void AvgAggregationFunction::reset(Nautilus::Value<Nautilus::MemRef> memref) {
     auto zero = Nautilus::Value<Nautilus::Int64>((int64_t) 0);
-    auto countMemref = (memref+offsetof(AvgAggregationValue, count)).as<Nautilus::MemRef>();
+    auto countMemref = (memref+ (uint64_t) offsetof(AvgAggregationValue, count)).as<Nautilus::MemRef>();
 
     memref.store(zero);
     countMemref.store(zero);
